+++ bin/language-export.ts
@@ -39,8 +39,8 @@ async function main() {
   if (!lang) {
     throw new Error(`no such language ${lang}`);
   }
-  const english = readJSON("../public/locales/en-translation.json");
-  const other = readJSON(`../public/locales/${lang}-translation.json`) || {};
+  const english = readJSON("../public/locales/en.json");
+  const other = readJSON(`../public/locales/${lang}.json`) || {};
   const csvData = [["Key", "en", lang], ...walk({ english, other })];
   const csv = Papa.unparse(csvData, { header: true });
   fs.writeFileSync(path.resolve(__dirname, `../${lang}.csv`), csv, "utf-8");