+++ src/i18n.ts
@@ -16,6 +16,7 @@ export const i18n = i18next
       escapeValue: false,
     },
     backend: {
-      loadPath: "/locales/{{lng}}-{{ns}}.json",
+      // loadPath: "/locales/{{lng}}-{{ns}}.json",
+      loadPath: "/locales/{{lng}}.json",
     },
   });